several speed improvements

+++ build.sh
@@ -1,9 +1,21 @@
 #!/bin/bash
 set -e
 
+IMAGE_NAME=${IMAGE_NAME:-$1}
+
+run_hook() {
+	HOOK=hooks/$1
+	if [ -f "$HOOK" ]; then
+		echo "=> Executing $HOOK hook"
+		./$HOOK
+	fi
+}
+
+EXTERNAL_DOCKER=no
 if [ -S /var/run/docker.sock ]; then
 	echo "=> Detected unix socket at /var/run/docker.sock"
 	docker version > /dev/null 2>&1 || (echo "   Failed to connect to docker daemon at /var/run/docker.sock" && exit 1)
+	EXTERNAL_DOCKER=yes
 else
 	echo "=> Starting docker"
 	wrapdocker > /dev/null 2>&1 &
@@ -47,24 +59,39 @@ if [ ! -d /app ]; then
 		echo "   ERROR: No application found in /app, and no \$GIT_REPO defined"
 		exit 1
 	fi
+	run_hook post_checkout
 else
 	echo "   Using existing app in /app"
 	cd /app
 fi
-cd .${DOCKERFILE_PATH-/}
+cd .${DOCKERFILE_PATH:-/}
+if [ -d "hooks" ]; then
+	chmod +x hooks/*
+fi
 
 if [ ! -f Dockerfile ]; then
 	echo "   WARNING: no Dockerfile detected! Created one using tutum/buildstep"
 	echo "FROM tutum/buildstep" >> Dockerfile
 fi
 
+echo "=> Building repository"
+run_hook pre_build
+docker build --rm --force-rm -t app_sut_1 .
+run_hook post_build
+
 echo "=> Testing repo"
-TEST_FILENAME=${TEST_FILENAME-docker-compose-test.yml}
+TEST_FILENAME=${TEST_FILENAME:-docker-compose.test.yml}
+if [ ! -f "./${TEST_FILENAME}" ] && [ -f "./docker-compose-test.yml" ]; then
+	echo "   WARNING: docker-compose-test.yml is deprecated. Rename your test file to docker-compose.test.yml"
+	TEST_FILENAME=docker-compose-test.yml
+fi
+
+run_hook pre_test
 if [ -f "./${TEST_FILENAME}" ]; then
-	# Next command is to workaround the fact that docker-compose does not use .dockercfg to pull images
-	# TODO: remove when fixed
-	cat ./${TEST_FILENAME} | grep "image:" | awk '{print $2}' | xargs -n1 docker pull
 	echo "=>  Executing tests"
+	# Next command is to workaround the fact that docker-compose does not use .dockercfg to pull images
+	# cat ./${TEST_FILENAME} | grep "image:" | awk '{print $2}' | xargs -n1 docker pull
+	docker-compose -f ${TEST_FILENAME} -p app build sut
 	docker-compose -f ${TEST_FILENAME} -p app up sut
 	RET=$(docker wait app_sut_1)
 	docker-compose -f ${TEST_FILENAME} -p app kill
@@ -78,16 +105,21 @@ if [ -f "./${TEST_FILENAME}" ]; then
 else
 	echo "   No tests found - skipping (have you created a ${TEST_FILENAME} file?)"
 fi
+run_hook post_test
 
 if [ ! -z "$IMAGE_NAME" ]; then
-	echo "=> Building image $IMAGE_NAME"
-	docker build --rm --force-rm -t $IMAGE_NAME .
 	if [ ! -z "$USERNAME" ] || [ ! -z "$DOCKERCFG" ] || [ -f /.dockercfg ]; then
 		echo "=>  Pushing image $IMAGE_NAME"
+		run_hook pre_push
+		docker tag -f app_sut_1 $IMAGE_NAME
 		docker push $IMAGE_NAME
+		run_hook post_push
 		echo "=>  Pushed image $IMAGE_NAME"
-		docker rmi -f $(docker images -q --no-trunc -a) > /dev/null 2>&1 || true
+		if [ "$EXTERNAL_DOCKER" == "no" ]; then
+			echo "=>  Cleaning up images"
+			docker rmi -f $(docker images -q --no-trunc -a) > /dev/null 2>&1 || true
+		fi
 	fi
 else
-	echo "   Skipping build and push"
+	echo "   Skipping push"
 fi
